classes/autotools.bbclass: Dont bail out if gettext/config.rpath does not exist.
* Add a note if this file does not exist. The reason is that gettext can
  needs libiconv and libiconv needs gettext so we need to solve this
  catch 22

Acked-by: Martin Jansa <Martin.Jansa@gmail.com>

+++ classes/autotools.bbclass
@@ -125,7 +125,11 @@ autotools_do_configure() {
 			    echo "no" | glib-gettextize --force --copy
 			  fi
 			else if grep "^[[:space:]]*AM_GNU_GETTEXT" $CONFIGURE_AC >/dev/null; then
-			  cp ${STAGING_DATADIR}/gettext/config.rpath ${S}/
+			  if [ -e ${STAGING_DATADIR}/gettext/config.rpath ]; then
+			    cp ${STAGING_DATADIR}/gettext/config.rpath ${S}/
+			  else
+			    oenote ${STAGING_DATADIR}/gettext/config.rpath not found. gettext is not installed.
+			  fi
 			fi
 
 			fi
